requirements.txt
langchain 
langchain-openai 
langsmith
langchain[openai]
python-dotenv
streamlit 
openai 
faiss-cpu 
pypdf
pdfplumber 
langchain-community